+++ src/rapidfuzz/_utils.py
@@ -3,9 +3,8 @@
 
 from __future__ import annotations
 
-from typing import Any, Callable
-
 from enum import Flag
+from typing import Any, Callable
 
 
 class ScorerFlag(Flag):